+++ fileio.c
@@ -10,18 +10,18 @@ FILE *fs_open(char *path, uint32_t mode) {
     0);
 }
 
-int fs_seek(FILE *stream, off_t offset, int whence) {
-  return SetFilePointer(stream, offset, 0, whence);
+int fs_seek(FILE *stream, long offset, int whence) {
+  return (int) SetFilePointer(stream, offset, 0, (uint32_t) whence);
 }
 
-off_t fs_read(void *ptr, uint32_t size, uint32_t j, FILE *stream) {
+off_t fs_read(void *ptr, size_t size, size_t j, FILE *stream) {
   DWORD q;
-  return ReadFile(stream, ptr, size * j, &q, 0);
+  return ReadFile(stream, ptr, (uint32_t) (size * j), &q, 0);
 }
 
-off_t fs_write(const void *ptr, uint32_t size, uint32_t j, FILE *stream) {
+off_t fs_write(const void *ptr, size_t size, size_t j, FILE *stream) {
   DWORD q;
-  return WriteFile(stream, ptr, size * j, &q, 0);
+  return WriteFile(stream, ptr, (uint32_t) (size * j), &q, 0);
 }
 
 int fs_close(FILE *file) {
